feat: adds create and update methods to albums CRUD

+++ dti-digital-challenge-server/src/domain/useCases/albums/deleteAlbumUseCase.ts
@@ -1,10 +1,16 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { AlbumsService } from 'src/domain/services/albums.service';
 
 @Injectable()
 export class DeleteAlbumUseCase {
   constructor(private albumsService: AlbumsService) {}
   async execute(albumId: number) {
+    const album = await this.albumsService.checkAlbumExists(albumId);
+
+    if (!album) {
+      throw new NotFoundException('Album not found.');
+    }
+
     await this.albumsService.deleteAlbum(albumId);
   }
 }
